+++ Runtime/Controllers/DSC_Dialogue_CanvasGroupController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace DSC.DialogueSystem
 {
-    public class DSC_Dialogue_CanvasGroupController : MonoBehaviour
+    public class DSC_Dialogue_CanvasGroupController : BaseDialogueUIGroup
     {
         #region Variable
 
@@ -15,6 +15,12 @@ namespace DSC.DialogueSystem
         [SerializeField] DSC_Dialogue_CanvasController[] m_arrCanvasController;
 
 #pragma warning restore 0649
+        #endregion
+
+        #region Variable - Property
+
+        public override UIType GroupType { get { return UIType.Canvas; } }
+
         #endregion
 
         #endregion
@@ -42,19 +48,28 @@ namespace DSC.DialogueSystem
 
         #endregion
 
+        #region Base - Override
+
+        public override void SetEnable(int nIndex, bool bEnable)
+        {
+            var hController = GetCanvasController(nIndex);
+            hController?.SetEnable(bEnable);
+        }
+
+        public override void SetAllEnable(bool bEnable)
+        {
+            if (!HasCanvasInArray())
+                return;
+
+            for (int i = 0; i < m_arrCanvasController.Length; i++)
+            {
+                m_arrCanvasController[i]?.SetEnable(bEnable);
+            }
+        }
+
+        #endregion
+
         #region Main
-
-        public void ShowCanvas(int nIndex)
-        {
-            var hController = GetCanvasController(nIndex);
-            hController?.ShowCanvas();
-        }
-
-        public void HideCanvas(int nIndex)
-        {
-            var hController = GetCanvasController(nIndex);
-            hController?.HideCanvas();
-        }
 
         public void SetEnable(int nIndex,UIType eGroupType, bool bEnable)
         {
@@ -65,6 +80,11 @@ namespace DSC.DialogueSystem
         #endregion
 
         #region Helper
+
+        protected bool HasCanvasInArray()
+        {
+            return (m_arrCanvasController != null && m_arrCanvasController.Length > 0);
+        }
 
         protected DSC_Dialogue_CanvasController GetCanvasController(int nIndex)
         {