+++ Runtime/Controllers/DSC_Dialogue_RawImageGroupController.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace DSC.DialogueSystem
 {
-    public class DSC_Dialogue_ImageGroupController : BaseDialogueUIGroup
+    public class DSC_Dialogue_RawImageGroupController : BaseDialogueUIGroup
     {
         #region Variable
 
@@ -12,14 +12,14 @@ namespace DSC.DialogueSystem
 #pragma warning disable 0649
 
         [SerializeField] protected DSC_Dialogue_DataController m_hDataController;
-        [SerializeField] protected DSC_Dialogue_ImageController[] m_arrImageController;
+        [SerializeField] protected DSC_Dialogue_RawImageController[] m_arrRawImageController;
 
 #pragma warning restore 0649
         #endregion
 
         #region Variable - Property
 
-        public override UIType GroupType { get { return UIType.Image; } }
+        public override UIType GroupType { get { return UIType.RawImage; } }
 
         #endregion
 
@@ -31,7 +31,7 @@ namespace DSC.DialogueSystem
         {
             if (m_hDataController && m_hDataController.DialogueEventDataList != null)
             {
-                m_hDataController.DialogueEventDataList.Add(new DialogueEventData_Image
+                m_hDataController.DialogueEventDataList.Add(new DialogueEventData_RawImage
                 {
                     m_hGroupController = this,
                 });
@@ -42,7 +42,7 @@ namespace DSC.DialogueSystem
         {
             if (m_hDataController != null && m_hDataController.DialogueEventDataList != null)
             {
-                m_hDataController.DialogueEventDataList.Remove<DialogueEventData_Image>();
+                m_hDataController.DialogueEventDataList.Remove<DialogueEventData_RawImage>();
             }
         }
 
@@ -52,22 +52,18 @@ namespace DSC.DialogueSystem
 
         public override void SetEnable(int nIndex, bool bEnable)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex);
             hController?.SetEnable(bEnable);
         }
 
         public override void SetAllEnable(bool bEnable)
         {
-            if (!HasImageInArray())
+            if (!HasRawImageInArray())
                 return;
 
-            for (int i = 0; i < m_arrImageController.Length; i++)
+            for(int i = 0; i < m_arrRawImageController.Length; i++)
             {
-                var hImageController = m_arrImageController[i];
-                if (hImageController == null)
-                    continue;
-
-                hImageController.SetEnable(bEnable);
+                m_arrRawImageController[i]?.SetEnable(bEnable);
             }
         }
 
@@ -75,66 +71,66 @@ namespace DSC.DialogueSystem
 
         #region Main
 
-        public void SetSprite(int nIndex, Sprite hSprite)
+        public void SetTexture(int nIndex, Texture hTexture)
         {
-            var hController = GetImageController(nIndex);
-            hController?.SetImage(hSprite);
+            var hController = GetRawImageController(nIndex);
+            hController?.SetImage(hTexture);
         }
 
         public void SetPosition(int nIndex, Vector2 vPosition)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex);
             hController?.SetPosition(vPosition);
         }
 
         public void SetSize(int nIndex, Vector2 vSize)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex);
             hController?.SetSize(vSize);
         }
 
         public void SetSizeToNative(int nIndex)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex);
             hController?.SetSizeToNative();
         }
 
-        public void SetRotation(int nIndex,Vector3 vRotation)
+        public void SetRotation(int nIndex, Vector3 vRotation)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex);
             hController?.SetRotation(vRotation);
         }
 
-        public void SetColor(int nIndex,Color hColor)
+        public void SetColor(int nIndex, Color hColor)
         {
-            var hController = GetImageController(nIndex);
+            var hController = GetRawImageController(nIndex); ;
             hController?.SetColor(hColor);
         }
-        
+
         #endregion
 
         #region Helper
 
-        protected bool HasImageInArray()
+        protected bool HasRawImageInArray()
         {
-            return (m_arrImageController != null && m_arrImageController.Length > 0);
+            return (m_arrRawImageController != null && m_arrRawImageController.Length > 0);
         }
 
-        protected DSC_Dialogue_ImageController GetImageController(int nIndex)
+        protected DSC_Dialogue_RawImageController GetRawImageController(int nIndex)
         {
-            DSC_Dialogue_ImageController hResult = null;
+            DSC_Dialogue_RawImageController hResult = null;
 
-            if (nIndex < 0 || m_arrImageController == null ||  m_arrImageController.Length <= nIndex)
+            if (nIndex < 0 || m_arrRawImageController == null || m_arrRawImageController.Length <= nIndex)
                 return hResult;
 
-            hResult = m_arrImageController[nIndex];
-            
+            hResult = m_arrRawImageController[nIndex];
+
             return hResult;
         }
 
-        protected bool TryGetImageController(int nIndex,out DSC_Dialogue_ImageController hOutController)
+        protected bool TryGetRawImageController(int nIndex, out DSC_Dialogue_RawImageController hOutController)
         {
-            hOutController = GetImageController(nIndex);
+            hOutController = GetRawImageController(nIndex);
 
             return hOutController != null;
         }